pyproject.toml
[project]
name = "Nexus"
version = "0.1"
description = "A proprietary algorithmic trading system."
readme = "README.md" 
authors = [
    { name = "Jared Grossthe", email = "jaredgrossthe@gmail.com" }
]
license = { file = "LICENSE" }  # Link to your license file
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "alpaca-py==0.37.0",
    "annotated-types==0.7.0",
    "boto3==1.36.10",
    "botocore==1.36.10",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "flake8==7.1.1",
    "future==1.0.0",
    "idna==3.10",
    "iniconfig==2.0.0",
    "jmespath==1.0.1",
    "joblib==1.4.2",
    "mccabe==0.7.0",
    "msgpack==1.1.0",
    "nolds==0.6.1",
    "numpy==2.2.2",
    "packaging==24.2",
    "pandas==2.2.3",
    "patsy==1.0.1",
    "pluggy==1.5.0",
    "pycodestyle==2.12.1",
    "pydantic==2.10.6",
    "pydantic_core==2.27.2",
    "pyflakes==3.2.0",
    "pytest==8.3.4",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-gnupg==0.5.4",
    "python-json-logger==3.2.1",
    "pytz==2024.2",
    "requests==2.32.3",
    "s3transfer==0.11.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.1",
    "setuptools==75.8.0",
    "six==1.17.0",
    "sseclient-py==1.8.0",
    "statsmodels==0.14.4",
    "threadpoolctl==3.5.0",
    "typing_extensions==4.12.2",
    "tzdata==2025.1",
    "urllib3==2.3.0",
    "websockets==14.2",
    "wheel==0.45.1"
]
requires-python = ">=3.12"

[tool.setuptools.packages.find]
where = ["."] 